\name{create.panel.of.normal.dataframe}
\alias{create.panel.of.normal.dataframe}
\title{
Create the panel of normals (PON) dataframe.
}
\description{From a cohort of normal samples, create the PON mutation dataframe.}
\usage{
create.panel.of.normal.dataframe(files = NULL)
}
\arguments{
  \item{files}{character vector containing full path to the mutation data files to include in the panel of normal analysis}
}
\value{
Returns a dataframe containing the normal data that be can be filtered.
}
\author{
Richard J. de Borja <richard.deborja@sickkids.ca>
}
\keyword{ snvid }
\keyword{ pon }
