\name{get_cigar_type}
\alias{get_cigar_type}
\title{
get_cigar_type
}
\description{
  Produces the title/name of a CIGAR string operation based on a given character.
}
\usage{
get_cigar_type(operation = NULL)
}
\arguments{
  \item{operation}{
Character representing a CIGAR string operation, either M, I, D, N, S, H, P, =, or X.
}
}

\value{
Produces the full title/category of the inputted CIGAR string character.
}
\author{
L. Christine Schreiner
}
\keyword{ CIGAR }
\keyword{ cigar }
\keyword{ type }
\keyword{ title }
\keyword{ category }
